refactor: rename mod sync to mod synchronous for symmetry with mod asynchronous
